Fix out of range error + continue decode

+++ src/PRD/Result.cpp
@@ -8,18 +8,18 @@ Result::Result(Instance* instance)
 	Z = vector<bool>();
 	z = vector<vector<bool>>();
 	x = vector<vector<vector<bool>>>();
-	D_M = vector<int>();
+	D_M = vector<double>();
 	C = vector<vector<double>>();
-	F = vector<int>();
-	f = vector<int>();
-	PT_M = vector<unsigned int>();
+	F = vector<double>();
+	f = vector<double>();
+	PT_M = vector<double>();
 	IC_WIP = 0;
 	IC_FIN = 0;
 	IC = 0;
 	PPC_M = 0;
 
-	D_3PL = vector<int>();
-	T_M = vector<int>();
+	D_3PL = vector<double>();
+	T_M = vector<double>();
 	VC = 0;
 }
